+++ RELEASE_NOTES.md
@@ -393,3 +393,12 @@
 ## Version 1.5.3
 
 - Added CBOR decoding function `iot_data_from_cbor`
+
+## Version 1.5.4
+
+- Added function `iot_data_alloc_const_f32` to allocate a const float data value
+- Added function `iot_data_alloc_const_f64` to allocate a const double data value
+- Fixed reading of values which exceed the size of int64_t and uint64_t in `iot_data_primitive_from_json`
+- Support added for Ubuntu 24.04
+- Support added for Fedora 40 and removed for Fedora 36
+- Added functions `iot_data_restricted_element_type` and `iot_data_restrict_element` to restrict the element type of lists, vectors and maps